First try at a long temr passion project.
Hope it goes well.

Currently it works to build static html files with no header/etc.

This doesn't check for files already existing and will have overright
them which means it can loop forever.

This is going to hit pythons maximum recursion depth for any reasonably
sized book but i should be able to linearize it do it it with multiple
processess.